B-Spline Surface with WebGPU - 원을 나타내는 점 그리기 구현 완료
원을 나타내는 점 WebGPU를 통해 구현 완료.
일부 불필요한 코드 수정.

+++ 실험장/실험_b-spline Surface.py
@@ -53,8 +53,8 @@ def calB_Spline(cps, knts, degree, numJoints=30):
     # print(uDraws)
 
     # 그릴 점 (u, v) - (start <= u, v <= end)     ## 원 -> 임의의 크기를 정해서 비율을 줄임
-    uDraws = [int(500 + 400 * math.cos(math.radians(theta))) / 1000 * (domainNum - 1) + start for theta in range(0, 372, 12)]
-    vDraws = [int(500 + 400 * math.sin(math.radians(theta))) / 1000 * (domainNum - 1) + start for theta in range(0, 372, 12)]
+    uDraws = [(500 + 400 * math.cos(math.radians(theta))) / 1000 * (domainNum - 1) + start for theta in range(0, 372, 12)]
+    vDraws = [(500 + 400 * math.sin(math.radians(theta))) / 1000 * (domainNum - 1) + start for theta in range(0, 372, 12)]
     print(uDraws)
     print(vDraws)
